refactor: write-config/read-configモジュールを統合
