+++ lib/config-file-io.js
@@ -2,8 +2,8 @@ const fs = require('fs')
 const path = require('path')
 const moment = require('moment')
 const { CONFIG_DIR, TOKEN_FILE, CREDENTIALS_FILE } = require('./constants')
-const readConfig = require('./read-config')
 const { encrypt } = require('./encrypt')
+const { decrypt } = require('./encrypt')
 
 const EXPIRES_IN = 3600
 const calcExpiry = () =>
@@ -11,7 +11,13 @@ const calcExpiry = () =>
     .add(EXPIRES_IN, 'seconds')
     .format()
 
-module.exports = ({ domain, token, credentials, password }) => {
+const decryptCredentias = (encryptedCredentials, password) => {
+  const decreptedStr = decrypt(encryptedCredentials, password)
+  const decreptedObj = JSON.parse(decreptedStr)
+  return decreptedObj
+}
+
+const writeConfig = ({ domain, token, credentials, password }) => {
   // TODO: パスワードが指定されていなければ標準入力させる
   // TODO: tokenとcredentialsで関数を分けてそれぞれexports
 
@@ -28,3 +34,21 @@ module.exports = ({ domain, token, credentials, password }) => {
   const encryptedCredentials = encrypt(JSON.stringify(credentials), password)
   fs.writeFileSync(path.join(domainDir, CREDENTIALS_FILE), encryptedCredentials, 'utf-8')
 }
+
+const readConfig = ({ domain, password }) => {
+  // TODO: パスワードが指定されていなければ標準入力させる
+  // TODO: tokenとcredentialsで関数を分けてそれぞれexports
+
+  // 保存済みトークン
+  const tokenFilePath = path.join(CONFIG_DIR, domain, TOKEN_FILE)
+  const token = fs.existsSync(tokenFilePath) ? require(tokenFilePath) : {}
+
+  // 保存済みクレデンシャル
+  const credentialsFilePath = path.join(CONFIG_DIR, domain, CREDENTIALS_FILE)
+  const encryptedCredentials = fs.existsSync(credentialsFilePath) && fs.readFileSync(credentialsFilePath, 'utf-8')
+  const credentials = encryptedCredentials ? decryptCredentias(encryptedCredentials, password) : {}
+
+  return { token, credentials }
+}
+
+module.exports = { writeConfig, readConfig }
