+++ lib/main.js
@@ -1,8 +1,7 @@
 const { isPast } = require('date-fns')
-const readConfig = require('./read-config')
 const server = require('./server')
 const refresh = require('./refresh')
-const writeConfig = require('./write-config')
+const { writeConfig, readConfig } = require('./config-file-io')
 
 module.exports = async argv => {
   const { domain, password } = argv
